# Conflicts:
#	src/main/java/unl/edu/ec/fieldPal/model/TimeSlot.java
#	src/main/java/unl/edu/ec/fieldPal/model/User.java
#	src/main/java/unl/edu/ec/fieldPal/service/security/CourtService.java
#	src/main/java/unl/edu/ec/fieldPal/service/security/OrganizationService.java
#	src/main/java/unl/edu/ec/fieldPal/service/security/ReservationService.java
#	src/main/java/unl/edu/ec/fieldPal/service/security/ScheduleService.java

+++ src/main/java/unl/edu/ec/fieldPal/controller/GestionBean.java
@@ -21,6 +21,7 @@ import java.io.Serializable;
 import java.util.List;
 
 /**
+ * @author NeoCoreTeam
  * Managed Bean para el panel de administración
  */
 
@@ -86,7 +87,7 @@ public class GestionBean implements Serializable {
     public CourtType[] getCourtTypes() { return CourtType.values(); }
 
     // === Dashboard stats ===
-    public int getTotalCanchas() { return courtService.getCourtCount(); }
+    public int getTotalCanchas() { return Math.toIntExact(courtService.getCourtCount()); }
     public int getReservasActivas() { return reservationService.getActiveCount(); }
     public int getUsuarios() { return userService.getUserCount(); }
     public double getIngresosMes() { return reservationService.getMonthlyIncome(); }
